telegram-lambda/src/requirements.txt
# AWS SDK for Python
boto3>=1.34.0
botocore>=1.34.0

# Type hints support
typing-extensions>=4.9.0

# Telegram Bot API
python-telegram-bot==21.0.1

# HTTP requests for Telegram file download
requests>=2.31.0

# AWS Embedded Metrics Format
aws-embedded-metrics>=3.0.0
